+++ Engine/Yields.cs
@@ -14,10 +14,13 @@
 		/// <summary>
 		/// Called when an update event occurs.
 		/// </summary>
+		/// <param name="rumor">
+		/// The rumor using this yield.
+		/// </param>
 		/// <param name="delta">
 		/// The amount of time in seconds since the last time this was called.
 		/// </param>
-		public virtual void OnUpdate(float delta) { }
+		public virtual void OnUpdate(Rumor rumor, float delta) { }
 
 		/// <summary>
 		/// Called when an advance event occurs.
@@ -60,7 +63,7 @@
 		/// <summary>
 		/// The number of seconds left to make a choice.
 		/// </summary>
-		public float SecondsLeft { get { return seconds; } }
+		public float SecondsLeft { get { return secondsLeft; } }
 		public int Default { get { return @default; } }
 
 		public ForChoice(int number, float seconds, int @default)
@@ -72,8 +75,13 @@
 			this.doUpdate = seconds > 0;
 		}
 
-		public override void OnUpdate(float delta)
+		public override void OnUpdate(Rumor rumor, float delta)
 		{
+			// Return early if there are no choices left
+			if (rumor != null && rumor.State.Choices.Count == 0) {
+				Finished = true;
+			}
+
 			if (!doUpdate || Finished) {
 				return;
 			}
@@ -81,7 +89,13 @@
 			if (secondsLeft > 0) {
 				secondsLeft -= delta;
 			}
-			Finished = secondsLeft <= 0;
+
+			if (secondsLeft <= 0) {
+				if (rumor != null) {
+					rumor.Choose(@default);
+				}
+				Finished = true;
+			}
 		}
 
 		public override void OnChoice()
@@ -110,7 +124,7 @@
 			this.seconds = seconds;
 		}
 
-		public override void OnUpdate(float delta)
+		public override void OnUpdate(Rumor rumor, float delta)
 		{
 			if (seconds > 0) {
 				seconds -= delta;